# docs/智能内容运营平台/设计/设计细节/openapi.yaml
openapi: 3.0.3
info:
  title: Agentic Content Studio API
  version: 1.0.0
  description: 智能内容运营平台 API，覆盖账号、选题、内容、Agent、审核、发布和数据复盘主流程。
servers:
  - url: http://localhost:3002/api
    description: 本地 API 服务
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 20
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          nullable: true
...
    Platform:
      type: string
      enum: [WECHAT, XIAOHONGSHU, DOUYIN, VIDEO_CHANNEL, BILIBILI, ZHIHU, OTHER]
    ContentStatus:
      type: string
      enum: [DRAFT, PENDING_GENERATE, GENERATING, PENDING_REVIEW, REJECTED, APPROVED, PENDING_PUBLISH, PUBLISHING, PUBLISHED, FAILED, REVIEWED, ARCHIVED]
    AgentType:
      type: string
      enum: [TITLE, TAG, REWRITE, BODY, COVER_COPY, REVIEW, SUMMARY, TOPIC, IMAGE, VIDEO_SCRIPT, COMPETITOR]
    RunStatus:
      type: string
      enum: [PENDING, RUNNING, SUCCESS, FAILED]
...
    Content:
      type: object
      properties:
        id:
          type: string
        topicId:
          type: string
          nullable: true
        title:
          type: string
        summary:
          type: string
          nullable: true
        body:
          type: string
          nullable: true
        coverText:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/ContentStatus'
        aiGenerated:
          type: boolean
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ContentVersion:
      type: object
      properties:
        id:
          type: string
        contentId:
          type: string
        platform:
          $ref: '#/components/schemas/Platform'
        accountId:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        body:
          type: string
          nullable: true
        coverText:
          type: string
          nullable: true
        tags:
          type: array
          nullable: true
          items:
            type: string
        formatConfig:
          type: object
          nullable: true
          additionalProperties: true
        status:
          $ref: '#/components/schemas/ContentStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Prompt:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        agentType:
          $ref: '#/components/schemas/AgentType'
        version:
          type: string
        template:
          type: string
        variables:
          type: object
          nullable: true
          additionalProperties: true
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PromptCreate:
      type: object
      required: [name, agentType, template]
      properties:
        name:
          type: string
        agentType:
          $ref: '#/components/schemas/AgentType'
        template:
          type: string
        version:
          type: string
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          $ref: '#/components/schemas/AgentType'
        description:
          type: string
          nullable: true
        promptId:
          type: string
        model:
          type: string
        enabled:
          type: boolean
        config:
          type: object
          nullable: true
          additionalProperties: true
        prompt:
          $ref: '#/components/schemas/Prompt'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AgentRun:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
        contentId:
          type: string
        versionId:
          type: string
          nullable: true
        input:
          type: object
          nullable: true
          additionalProperties: true
        output:
          type: object
          nullable: true
          additionalProperties: true
        model:
          type: string
          nullable: true
        promptVersion:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/RunStatus'
        error:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        durationMs:
          type: integer
          nullable: true
          description: 执行耗时（毫秒），规划字段，V1.1 实现
        agent:
          $ref: '#/components/schemas/Agent'
        content:
          $ref: '#/components/schemas/Content'
        version:
          $ref: '#/components/schemas/ContentVersion'
...
    ImaConfigUpdate:
      type: object
      properties:
        clientId:
          type: string
        apiKey:
          type: string
        baseUrl:
          type: string
        useMock:
          type: boolean
...
    ImaSearch:
      type: object
      required: [query, contentId]
      properties:
        query:
          type: string
        contentId:
          type: string
        limit:
          type: integer
        knowledgeBaseId:
          type: string
...
paths:
...
  /ima/config:
    get:
      tags: [IMA]
      summary: IMA 配置
      responses:
        '200':
          description: OK
    put:
      tags: [IMA]
      summary: 更新 IMA 配置
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImaConfigUpdate'
      responses:
        '200':
          description: OK
  /ima/knowledge-bases:
    get:
      tags: [IMA]
      summary: 知识库列表
      parameters:
        - name: enabledOnly
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: OK
...
  /ima/knowledge-bases/{id}:
    patch:
      tags: [IMA]
      summary: 更新知识库配置
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
                isDefault:
                  type: boolean
                name:
                  type: string
      responses:
        '200':
          description: OK
  /ima/search:
    post:
      tags: [IMA]
      summary: IMA 搜索
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImaSearch'
      responses:
        '200':
          description: OK
...
  /agent-runs:
    get:
      tags: [AgentRuns]
      summary: Agent 执行记录
      x-implementation-status: implemented
      parameters:
        - name: contentId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/RunStatus'
        - name: agentType
          in: query
          schema:
            $ref: '#/components/schemas/AgentType'
          description: V1.1 规划
        - name: platform
          in: query
          schema:
            $ref: '#/components/schemas/Platform'
          description: V1.1 规划
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
          description: V1.1 规划
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
          description: V1.1 规划
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AgentRun'
  /agent-runs/{id}:
    get:
      tags: [AgentRuns]
      summary: Agent 执行详情
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AgentRun'
...
  /prompts:
    get:
      tags: [Prompts]
      summary: Prompt 列表
      responses:
        '200':
          description: OK
    post:
      tags: [Prompts]
      summary: 创建 Prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromptCreate'
      responses:
        '200':
          description: OK
  /prompts/{id}:
    patch:
      tags: [Prompts]
      summary: 更新 Prompt
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: OK
...